rename changeCombinator, separate tests
