+++ packages/dds/tree/src/test/sequence-change-family/randomSequenceGenerator.spec.ts
@@ -4,33 +4,10 @@
  */
 
 import { strict as assert } from "assert";
-import { makeRandom } from "@fluid-internal/stochastic-test-utils";
-import { ChangeRebaser } from "../../rebase";
-import { AnchorSet, FieldKey } from "../../tree";
-import { brand, changeCombinator, generateRandomChange, generateRandomUpPaths } from "../../util";
+import { FieldKey } from "../../tree";
+import { brand, generateRandomChange, generateRandomUpPaths } from "../../util";
 
 const testSeed = 432167897;
-
-function commutativeRebaser<TChange>(data: {
-    compose: (changes: TChange[]) => TChange;
-    invert: (changes: TChange) => TChange;
-    rebaseAnchors: (anchor: AnchorSet, over: TChange) => void;
-}): ChangeRebaser<TChange> {
-    return {
-        rebase: (change: TChange, over: TChange) => change,
-        ...data,
-    };
-}
-
-const counterRebaser = commutativeRebaser({
-    compose: (changes: number[]) => changes.reduce((a, b) => a + b, 0),
-    invert: (change: number) => -change,
-    rebaseAnchors: (anchor: AnchorSet, over: number) => {},
-});
-
-function generateRandomCounterChange(seed: number) {
-    return makeRandom(seed).integer(-1000, 1000);
-}
 
 describe("Test randomSequenceBuilder", () => {
     it("consistency of generateRandomUpPaths with same seed.", () => {
@@ -45,21 +22,6 @@ describe("Test randomSequenceBuilder", () => {
 
         const change1 = generateRandomChange(upPaths, testSeed);
         const change2 = generateRandomChange(upPaths, testSeed);
-        assert.deepStrictEqual(change1, change2);
-    });
-    it("consistency of the changeCombinator using counter field kind.", () => {
-        const change1 = changeCombinator(
-            counterRebaser,
-            (seed) => generateRandomCounterChange(seed),
-            testSeed,
-            10,
-        );
-        const change2 = changeCombinator(
-            counterRebaser,
-            (seed) => generateRandomCounterChange(seed),
-            testSeed,
-            10,
-        );
         assert.deepStrictEqual(change1, change2);
     });
 });
